Use case 2: Using Google Guice to provide method security
Key features:
 - the existing user / roles / department / status model is reused (it is extended for rights,  that could be unique for each operation -- so one can disable adding courses for instance)
 - implemented using Guice and its AOP capabilities (using method interceptors)
 - security can be provided on methods (using annotations, like  @CheckForPermission(permission = {ClassEditPermission.class})) or programmatically using injections
  - this can be easily extended by providing JSP tags or on URLS using a filter or filters
 - security can be based on roles (has role curriculum manager),  rights (role has a set of rights, e.g., class-edit), or permissions (custom code checking particular object -- see ClassEditPermission)
   - roles can be hierarchical (defined using inheritance)
 - see CurriculaServlet for an example on how to secure various methods
 
 So for example, 
 
   @CheckDepartment(right = {Right.CurriculumAdd, Right.CurriculumEdit}, all = true)
   void proceed(Long departmentId, ...) { ... }

means that the user has to be an administrator or has to have curriculum-add and curriculum-edit rights, and when the role is derived from DepartmentDependentRole (like departmental schedule manager), the user also has to be associated with the given department (passed as method argument, recognized either by the type (Department class), or by the first Long argument for the department uniqueId)

Another example:
 
  @CheckPermission(permission = {ClassEditPermission.class})
   void save(Class_ clazz) { ... }

In this case the ClassEditPermission.check(Class_ clazz) is invoked, including user context (user's role, department(s), current academic session etc) -- so that the whole business of checking controlling or managing department against department (or session) status can be done.

Last example:

  @Inject Provider<CurriculumPermissions.CanDelete> permission;
  public someUnsecuredDeleteMethod(List<Curriculum> curricula) {
    for (Curriculum c: curricula)
      if (permission.get().check(c)) hibSession.delete(c); 
  }

A particular permission check can be programmatically invoked using an injected permission.

The whole think can be easily customized injecting custom implementation, for instance:

		bind(CurriculumPermissions.CanEdit.class).to(CurriculumEditAlwaysDenyPermission.class);
		bind(CurriculumMgrRole.class).to(CurriculumMgrRoleNoAddDelete.class);